application/pyproject.toml
[tool.poetry]
authors = ["shun198"]
description = "api"
name = "api"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
Authlib = "^1.1.0"
Django = "^4.2.3"
boto3 = "^1.26.22"
celery = "^5.2.7"
chardet = "^5.2.0"
django-celery-beat = "^2.4.0"
django-cors-headers = "^3.13.0"
django-filter = "^22.1"
django-ses = "^3.2.2"
django-storages = "^1.13.1"
djangorestframework = "^3.14.0"
drf-nested-routers = "^0.93.4"
drf-spectacular = "^0.24.2"
gunicorn = "^20.1.0"
kombu = "^5.2.4"
pillow = "^9.4.0"
psycopg2 = "^2.9.6"
pycurl = "^7.45.2"
pydantic = "^1.10.6"
python = "^3.11"
redis = "^4.3.4"
requests = "^2.31.0"
toml = "^0.10.2"
graphene-django = "^3.2.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
django-debug-toolbar = "^3.8.1"
django-extensions = "^3.2.1"
factory-boy = "^3.2.1"
freezegun = "^1.2.2"
isort = "^5.11.4"
pdoc = "^13.0.0"
pytest = "^7.1.3"
pytest-bdd = "^6.1.1"
pytest-clarity = "^1.0.1"
pytest-cov = "^4.0.0"
pytest-custom-exit-code = "^0.3.0"
pytest-django = "^4.5.2"
pytest-github-actions-annotate-failures = "^0.2.0"
pytest-mock = "^3.10.0"
pytest-sugar = "^0.9.6"
pytest-xdist = "^3.0.2"

[tool.isort]
line_length = 79
profile = "black"

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
include = '\.py$'
line-length = 79

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings.local"
addopts = "-vv --diff-width=60 -s --durations=0"

# logの設定
[logging]
version = 1
[logging.formatters.simple]
format = "[%(levelname)s] %(name)s - %(message)s "
[logging.handlers.applicationHandler]
class = "logging.StreamHandler"
formatter = "simple"
level = "INFO"
stream = "ext://sys.stdout"
[logging.handlers.errorHandler]
class = "logging.StreamHandler"
formatter = "simple"
level = "ERROR"
stream = "ext://sys.stdout"
[logging.loggers.application]
handlers = ["applicationHandler"]
level = "DEBUG"
propagate = "no"
[logging.loggers.emergency]
handlers = ["errorHandler"]
level = "DEBUG"
propagate = "no"
